don't blindly jump releases
Don't track the moving target Debian Stable because this renders a lot of demos
broken at the time Debian releases a new stable. A situation that can last for
months depending on the schedule and availability of the developer of a
particular demo. During this time the demo can no longer be showed to or used by
third parties that might be interested in the project.

Instead, track the latest stable that is known to work and as soon as Debian
releases a new stable test all demos with it and coordinate this with the
individual owners of each demo before bumping the base image.

For now track Debian Stretch because that is the latest stable that is known to
work and thus supported by all of us. Since Buster has been released this
summer, work should be initiated to make sure all demos work with Buster so that
we can move the base image to it.